bluetooth_classic_stash/host_requirements.txt
pytest
utf_core
gcovr
scapy
jinja2
pyyaml
numpy
xmltodict